requirements.txt
flask
flask_navigation
tweepy
wordcloud
numpy 
matplotlib
re
nltk.corpus
PIL
requests